requirements.txt
Django==4.2.13
djangorestframework==3.15.2
psycopg2==2.9.9
transformers==4.42.3
torch==2.0.0
